Added find_point method to Lapper
Why: For Lapper<u32,T>, It is strange to have the
find mechanism be non_inclusive when searching for a
single point as it would make sense for the ends
to be included.

Critically, this implemention is for inclusive intervals.
It might be worth it to create a non_inclusive find_point() in
the future.

How: Refactored IterFind to include a FindQuery parameter,
also added contains_inclusive(point:I) to Interval